Admin emails

Closes #57 and #60

See merge request elixire/elixire!40

Former-commit-id: 01c3483006743b544fa3e1d9b18ecbdf49f74c23

+++ api/bp/admin/object.py
@@ -11,52 +11,53 @@ from api.schema import validate, ADMIN_MODIFY_FILE
 from api.errors import BadInput
 from api.decorators import admin_route
 
+from api.bp.admin.audit_log_actions.object import (
+    ObjectEditCtx, ObjectDeleteCtx
+)
+
+from api.common.fetch import (
+    OBJ_MAPPING
+)
+
 bp = Blueprint(__name__)
 
 
-async def generic_namefetch(table, request, shortname):
-    """Generic function to fetch a file or shorten
-    information based on shortname."""
-    fields = 'file_id, mimetype, filename, file_size, uploader, fspath, deleted, domain' \
-             if table == 'files' else \
-             'shorten_id, filename, redirto, uploader, deleted, domain'
+async def _handler_object(request, obj_type: str,
+                          obj_fname: str) -> response:
+    """Handler for fetching files/shortens."""
+    id_handler, obj_handler = OBJ_MAPPING[obj_type]
 
-    id_field = 'file_id' if table == 'files' else 'shorten_id'
+    conn = request.app.db
 
-    row = await request.app.db.fetchrow(f"""
-    SELECT {fields}
-    FROM {table}
-    WHERE filename = $1
-    """, shortname)
+    obj_id = await id_handler(conn, obj_fname)
 
-    if not row:
-        return response.json(None)
+    if obj_id is None:
+        raise BadInput('Object not found')
 
-    drow = dict(row)
-    drow[id_field] = str(drow[id_field])
-    drow['uploader'] = str(drow['uploader'])
-
-    return response.json(drow)
+    return response.json(
+        await obj_handler(conn, obj_id)
+    )
 
 
 @bp.get('/api/admin/file/<shortname>')
 @admin_route
-async def get_file_by_name(request, admin_id, shortname):
+async def get_file_by_name(request, _admin_id, shortname):
     """Get a file's information by shortname."""
-    return await generic_namefetch('files', request, shortname)
+    return await _handler_object(request, 'file', shortname)
 
 
 @bp.get('/api/admin/shorten/<shortname>')
 @admin_route
-async def get_shorten_by_name(request, admin_id, shortname):
+async def get_shorten_by_name(request, _admin_id, shortname):
     """Get a shorten's information by shortname."""
-    return await generic_namefetch('shortens', request, shortname)
+    return await _handler_object(request, 'shorten', shortname)
 
 
-async def handle_modify(atype: str, request, thing_id):
+async def handle_modify(obj_type: str, request,
+                        obj_id: int) -> response:
     """Generic function to work with files OR shortens."""
-    table = 'files' if atype == 'file' else 'shortens'
-    field = 'file_id' if atype == 'file' else 'shorten_id'
+    table = 'files' if obj_type == 'file' else 'shortens'
+    field = 'file_id' if obj_type == 'file' else 'shorten_id'
 
     payload = validate(request.json, ADMIN_MODIFY_FILE)
 
@@ -69,9 +70,9 @@ async def handle_modify(atype: str, request, thing_id):
     SELECT filename, domain
     FROM {table}
     WHERE {field} = $1
-    """, thing_id)
+    """, obj_id)
 
-    thing_name = row['filename']
+    obj_name = row['filename']
     old_domain = row['domain']
 
     if new_domain is not None:
@@ -80,14 +81,13 @@ async def handle_modify(atype: str, request, thing_id):
             UPDATE {table}
             SET domain = $1
             WHERE {field} = $2
-            """, new_domain, thing_id)
+            """, new_domain, obj_id)
         except asyncpg.ForeignKeyViolationError:
             raise BadInput('Unknown domain ID')
 
         # Invalidate based on the query
-        to_invalidate = f'fspath:{old_domain}:{thing_name}' \
-                        if atype == 'file' else \
-                        f'redir:{old_domain}:{thing_name}'
+        to_invalidate = 'fspath' if obj_type == 'file' else 'redir'
+        to_invalidate = f'{to_invalidate}:{old_domain}:{obj_name}'
 
         await request.app.storage.raw_invalidate(to_invalidate)
         updated.append('domain')
@@ -99,13 +99,13 @@ async def handle_modify(atype: str, request, thing_id):
             UPDATE {table}
             SET filename = $1
             WHERE {field} = $2
-            """, new_shortname, thing_id)
+            """, new_shortname, obj_id)
         except asyncpg.UniqueViolationError:
             raise BadInput('Shortname already exists.')
 
         # Invalidate both old and new
         await request.app.storage.raw_invalidate(*[
-            f'fspath:{old_domain}:{thing_name}',
+            f'fspath:{old_domain}:{obj_name}',
             f'fspath:{old_domain}:{new_shortname}',
         ])
 
@@ -116,21 +116,25 @@ async def handle_modify(atype: str, request, thing_id):
 
 @bp.patch('/api/admin/file/<file_id:int>')
 @admin_route
-async def modify_file(request, admin_id, file_id):
+async def modify_file(request, _admin_id, file_id):
     """Modify file information."""
-    return await handle_modify('file', request, file_id)
+
+    async with ObjectEditCtx(request, file_id, 'file'):
+        return await handle_modify('file', request, file_id)
 
 
 @bp.patch('/api/admin/shorten/<shorten_id:int>')
 @admin_route
-async def modify_shorten(request, admin_id, shorten_id):
+async def modify_shorten(request, _admin_id, shorten_id):
     """Modify file information."""
-    return await handle_modify('shorten', request, shorten_id)
+
+    async with ObjectEditCtx(request, shorten_id, 'shorten'):
+        return await handle_modify('shorten', request, shorten_id)
 
 
 @bp.delete('/api/admin/file/<file_id:int>')
 @admin_route
-async def delete_file_handler(request, admin_id, file_id):
+async def delete_file_handler(request, _admin_id, file_id):
     """Delete a file."""
     row = await request.app.db.fetchrow("""
     SELECT filename, uploader
@@ -138,7 +142,11 @@ async def delete_file_handler(request, admin_id, file_id):
     WHERE file_id = $1
     """, file_id)
 
-    await delete_file(request.app, row['filename'], row['uploader'])
+    if row is None:
+        raise BadInput('File not found')
+
+    async with ObjectDeleteCtx(request, file_id, 'file'):
+        await delete_file(request.app, row['filename'], row['uploader'])
 
     return response.json({
         'shortname': row['filename'],
@@ -149,7 +157,7 @@ async def delete_file_handler(request, admin_id, file_id):
 
 @bp.delete('/api/admin/shorten/<shorten_id:int>')
 @admin_route
-async def delete_shorten_handler(request, admin_id, shorten_id: int):
+async def delete_shorten_handler(request, _admin_id, shorten_id: int):
     """Delete a shorten."""
     row = await request.app.db.fetchrow("""
     SELECT filename, uploader
@@ -157,7 +165,13 @@ async def delete_shorten_handler(request, admin_id, shorten_id: int):
     WHERE shorten_id = $1
     """, shorten_id)
 
-    await delete_shorten(request.app, row['filename'], row['uploader'])
+    if row is None:
+        raise BadInput('Shorten not found')
+
+    async with ObjectDeleteCtx(request, shorten_id, 'shorten'):
+        await delete_shorten(request.app,
+                             row['filename'],
+                             row['uploader'])
 
     return response.json({
         'shortname': row['filename'],
